fix: Address Copilot code scanning findings for MCP module
Fixes GitHub Advanced Security (Copilot) findings:
1. Cyclic import: Extract CLI entry point to separate api/mcp/cli.py
   - api/mcp/server.py no longer imports api/main (breaking the cycle)
   - cli.py handles the import of api/main for CLI entry point only
   - Architecture: main.py -> server.py, cli.py -> main.py (no cycle)

2. Unused imports: Change test imports from class to module import
   - tests/unit/mcp/test_mcp_config.py: import pydantic_settings
   - tests/unit/mcp/test_mcp_server.py: import pydantic_settings

All checks pass: make check, make typecheck-fast, pytest tests/unit/mcp/
